fixing a bug with column set update
Column names in the flattened schema are stored as they are in the original dataset, that is with white space.
Columns that we get from the database are sanitized, i.e. with underscores instead of white space.
Because of this the update_columns method doesn't recognize that the schema columns already exist in the database.

This change adds sanitization for the column names in the flattened schema, so that we use the same format for column names in both sets.